refactor; identify function schedule by name
